fix(abi): correct extern fn param widths to match C (setSystemFlag, allocC47Blocks)
Extern-fn param-width scan across owners. Two functions were bound with a
narrower parameter than the C prototype:

- setSystemFlag: C is `void setSystemFlag(unsigned int)` but frontier_jm and
  frontier_integers declared `flag: u16`. Current flag encodings (0x8004..0xc0xx)
  fit u16, so it works today, but any flag >= 0x10000 would be silently truncated
  at the call. Widen to c_uint to match C.
- allocC47Blocks: C is `void *allocC47Blocks(size_t)` but frontier_addons declared
  `numberOfBlocks: u32`. Widen to usize.

Call-site values widen cleanly (u16/u32 -> c_uint/usize). Pointer/const/bool_t/
calcRegister_t spelling variants across owners are same-ABI and left as-is.

Verified: zig build sim and zig build test exit 0 (9674 testSuite), fmt clean.

+++ zig_src/frontier/frontier_addons.zig
@@ -650,7 +650,7 @@ inline fn freeRegisterData(regist: calcRegister_t) void {
     freeC47Blocks(getRegisterDataPointer(regist), getRegisterFullSizeInBlocks(regist));
 }
 extern fn reallocateRegister(regist: calcRegister_t, dataType: u32, dataSizeInBytes: u32, tag: u32) void;
-extern fn allocC47Blocks(numberOfBlocks: u32) [*c]u8;
+extern fn allocC47Blocks(numberOfBlocks: usize) [*c]u8;
 extern fn moreInfoOnError(m1: [*c]const u8, m2: [*c]const u8, m3: [*c]const u8, m4: [*c]const u8) void;
 extern fn fnDrop(unusedButMandatoryParameter: u16) void;
 extern fn fnDropY(unusedButMandatoryParameter: u16) void;
